데이터 요청 코드를 weeklyViewerRankingCard컴포넌트 안으로 옮김
